chore: moved back to /chainlock
