+++ .eslintignore
@@ -6,4 +6,4 @@ lib/crypto/bls.js
 lib/deterministicmnlist/SimplifiedMNList.js
 lib/deterministicmnlist/QuorumEntry.js
 lib/util/bitarray.js
-lib/messages/chainlock.js
+lib/chainlock/chainlock.js
