+++ index.d.ts
@@ -19,7 +19,7 @@ import * as _PartialMerkleTree from './typings/block/PartialMerkleTree';
 import * as _BufferWriter from './typings/buffer/BufferWriter';
 import * as _BufferReader from './typings/buffer/BufferReader';
 
-import * as _ChainLock from './typings/messages/ChainLock';
+import * as _ChainLock from './typings/chainlock/ChainLock';
 
 import * as _BN from './typings/crypto/BN';
 import * as _Point from './typings/crypto/Point';
