+++ lib/chainlock/chainlock.js
